services/api/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "card-reader-api"
version = "0.1.7"
description = "Card Reader API service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "card-reader-core",
  "django>=5.1.0",
  "djangorestframework>=3.15.2",
  "gunicorn>=23.0.0",
  "pydantic-settings>=2.5.2",
  "python-multipart>=0.0.12"
]

[tool.uv.sources]
card-reader-core = { path = "../core", editable = true }

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
include = ["card_reader_api*"]

[tool.setuptools.package-data]
card_reader_api = ["seeds/*.json", "seeds/assets/**/*"]

[project.optional-dependencies]
dev = [
  "mypy>=1.11.2",
  "pytest>=8.3.3",
  "httpx>=0.27.2",
  "ruff>=0.6.9",
  "watchfiles>=0.24.0"
]

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.12"
strict = true
mypy_path = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
